motion function to move object on the X and Y
