+++ Interactive-ervaring/Assets/Prefabs/TestPref/fishhook.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 346c9f5f9cfc97f41bb5cc86d85a0b56
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
