+++ Interactive-ervaring/Assets/Scenes/MobileMovementScene.unity
@@ -130,74 +130,89 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 5294364932301317698, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
-      propertyPath: m_Name
-      value: Crocodile
-      objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_RootOrder
       value: 3
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalScale.x
-      value: 1
+      value: 10
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalScale.y
-      value: 1
+      value: 10
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalScale.z
-      value: 1
+      value: 10
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 0
+      value: -1.58
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalPosition.z
       value: 3.21
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalRotation.w
-      value: 0
+      value: 0.4273195
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalRotation.x
-      value: 0
+      value: 0.4273195
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalRotation.y
-      value: 1
+      value: -0.56338096
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalRotation.z
-      value: 0
+      value: 0.56338096
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
+      value: 90
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
-      value: 180
+      value: -90
       objectReference: {fileID: 0}
-    - target: {fileID: 5398267037370213625, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: -8679921383154817045, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
+      value: 15.64
       objectReference: {fileID: 0}
-    - target: {fileID: 7041761704044929181, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
-      propertyPath: m_Enabled
-      value: 0
+    - target: {fileID: 919132149155446097, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
+      propertyPath: m_Name
+      value: FishhookLP
       objectReference: {fileID: 0}
-    m_RemovedComponents:
-    - {fileID: 7041761704044929181, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
-    - {fileID: 7041761704044929182, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
-  m_SourcePrefab: {fileID: 100100000, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
+    - target: {fileID: 919132149155446097, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
+      propertyPath: m_Layer
+      value: 5
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
+--- !u!1 &260227823 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: ca7cc2c672ef49f4dab326daa7104082, type: 3}
+  m_PrefabInstance: {fileID: 260227822}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &260227824
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 260227823}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ec7e7ce8cfb67ec47a6c632b1704b9f2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &304919557
 GameObject:
   m_ObjectHideFlags: 0
@@ -599,23 +614,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &1696746730 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 5294364932301317698, guid: 47a57ae5d8d090c46bdab2c299343dcf, type: 3}
-  m_PrefabInstance: {fileID: 260227822}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &1696746734
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1696746730}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ec7e7ce8cfb67ec47a6c632b1704b9f2, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1973090302
 GameObject:
   m_ObjectHideFlags: 0
